- no removal of items with swipe in shopping list page android! (unset them or remove through edit dlg) - update to qt 5.15.1 - compiler warnings

+++ librecipebook/src/data/Amount.cpp
@@ -51,7 +51,7 @@ float recipebook::Amount::getChangedAmount(float quantity, bool bIncrease)
 	}
 
 	const float* fPos = std::lower_bound(valueSteps, valueSteps + valueStepsCount, quantity);
-	int position = fPos - valueSteps;
+    int position = static_cast<int>(fPos - valueSteps);
     if(bIncrease)
 	{
 		if(position == valueStepsCount - 1)
